+++ Ui/Component/Listing/Column/MenuList/PageActions.php
@@ -21,6 +21,7 @@ class PageActions extends Column
      */
     const URL_PATH_EDIT = 'snowmenu/menu/edit';
     const URL_PATH_DELETE = 'snowmenu/menu/delete';
+    const URL_PATH_EXPORT = 'snowmenu/menu/export';
 
     /**
      * @var UrlInterface
@@ -64,6 +65,7 @@ class PageActions extends Column
                     $item[$name] = [
                         'edit' => $this->getEditButton($menuId),
                         'delete' => $this->getDeleteButton($menuId),
+                        'export' => $this->getExportButton($menuId)
                     ];
                 }
             }
@@ -110,6 +112,15 @@ class PageActions extends Column
             ],
             'post' => true,
             '__disableTmpl' => true,
+        ];
+    }
+
+    private function getExportButton(int $menuId): array
+    {
+        return [
+            'href' => $this->urlBuilder->getUrl(self::URL_PATH_EXPORT, ['id' => $menuId]),
+            'label' => __('Export'),
+            '__disableTmpl' => true
         ];
     }
 }